+++ init/router/router.go
@@ -1,0 +1,9 @@
+package router
+
+import "github.com/gofiber/fiber/v2"
+
+func InitRouter() *fiber.App {
+	app := fiber.New()
+
+	return app
+}