+++ internal/response/service/response_service.go
@@ -1,6 +1,9 @@
 package service
 
 import (
+	"net/url"
+	"strconv"
+
 	"github.com/gofiber/fiber/v2"
 
 	"github.com/ilyavovnenko/shops_categories_ms/internal/response"
@@ -9,8 +12,8 @@ import (
 /*
 * Responses
  */
-func SuccessResponse(c *fiber.Ctx, message string, data interface{}) error {
-	return c.JSON(getSuccessResponse(c, message, data))
+func SuccessResponse(c *fiber.Ctx, data interface{}) error {
+	return c.JSON(getSuccessResponse(c, data))
 }
 
 func InternalErrorResponse(c *fiber.Ctx, errors []response.AppError) error {
@@ -24,11 +27,10 @@ func BadRequestResponse(c *fiber.Ctx, errors []response.ValidationError) error {
 /*
 * Preparation block
  */
-func getSuccessResponse(c *fiber.Ctx, message string, data interface{}) response.Response {
+func getSuccessResponse(c *fiber.Ctx, data interface{}) response.Response {
 	return response.Response{
-		Data:    data,
-		Message: message,
-		Meta:    generateMeta(c),
+		Data: data,
+		Meta: generateMeta(c),
 	}
 }
 
@@ -54,10 +56,14 @@ func getValidationErrorResponse(c *fiber.Ctx, message string, errors []response.
 * META block
  */
 func generateMeta(c *fiber.Ctx) response.Meta {
+	parsedURL, _ := url.Parse(generateSelfLink(c))
+	currentPage := getCurrentPage(c)
+
 	return response.Meta{
-		Page:            0, // todo: add functionality for getting page and another staff
-		PerPage:         0,
-		Self:            generateSelfLink(c),
+		Page:            0,
+		Self:            parsedURL.String(),
+		Next:            generateNextLink(c, parsedURL, currentPage),
+		Previous:        generatePreviousLink(c, parsedURL, currentPage),
 		TotalPages:      0,
 		TotalItemsCount: 0,
 	}
@@ -65,4 +71,41 @@ func generateMeta(c *fiber.Ctx) response.Meta {
 
 func generateSelfLink(c *fiber.Ctx) string {
 	return c.BaseURL() + c.OriginalURL()
+}
+
+func getCurrentPage(c *fiber.Ctx) int {
+	numPage := c.Query("page")
+	currentPage, _ := strconv.Atoi(numPage)
+
+	if currentPage == 0 {
+		currentPage = 1
+	}
+
+	return currentPage
+}
+
+func generateNextLink(c *fiber.Ctx, parsedUrl *url.URL, currentPage int) string {
+	// TODO: check if the currentPage not bigger then last page
+
+	values, _ := url.ParseQuery(parsedUrl.RawQuery)
+	values.Set("page", strconv.Itoa(currentPage+1))
+
+	parsedUrl.RawQuery = values.Encode()
+
+	return parsedUrl.String()
+}
+
+func generatePreviousLink(c *fiber.Ctx, parsedUrl *url.URL, currentPage int) string {
+	prevPage := currentPage - 1
+
+	if prevPage < 1 {
+		return ""
+	}
+
+	values, _ := url.ParseQuery(parsedUrl.RawQuery)
+	values.Set("page", strconv.Itoa(prevPage))
+
+	parsedUrl.RawQuery = values.Encode()
+
+	return parsedUrl.String()
 }